Move monitoring logic to a controller concern
